feat: 폴더구조 정리 및 theme 적용
